Fix ratelimit
There is no good reason for the ratelimit helper function to use a separate thread. A simple sleep will do.
